ai-service/requirements.txt
groq==0.11.0
python-dotenv==1.0.1
flask==3.1.0
flask-limiter==3.8.0
sentence-transformers==3.0.1
chromadb==0.5.3
